requirements.txt
feedparser
PyYAML
openai
gtts
pydub
moviepy
Pillow
requests
